+++ sharc/output/inr.txt
@@ -1,60 +1,53 @@
- -48.00867
- -39.32241
- -44.85260
- -45.58674
- -47.79872
- -46.57525
- -48.07163
- -44.36399
- -46.69945
- -43.96878
- -48.59304
- -43.30336
- -39.03002
- -40.44205
- -43.62630
- -48.03996
- -43.55421
- -46.01866
- -46.47587
- -40.65495
- -46.20901
- -45.78941
- -46.18813
- -39.25616
- -48.19495
- -47.45865
- -49.03135
- -47.03703
- -46.00114
- -48.56236
- -46.78617
- -46.46143
- -47.96634
- -48.41746
- -40.07734
- -47.18763
- -47.31692
- -46.64947
- -41.48237
- -42.86452
- -43.59489
- -43.54702
- -39.39080
- -41.47975
- -46.14982
- -47.45104
- -41.81055
- -47.19233
- -43.12170
- -49.03290
- -41.84470
- -45.41817
- -45.46064
- -45.02902
- -45.54934
- -48.46374
- -43.79519
- -47.84720
- -51.52961
- -46.54972
+ -48.55398
+ -40.16666
+ -44.19292
+ -47.51876
+ -46.20785
+ -40.84643
+ -46.99232
+ -47.45599
+ -47.60627
+ -40.43999
+ -47.46435
+ -47.08493
+ -43.69930
+ -44.52520
+ -46.75915
+ -46.87056
+ -47.78132
+ -42.55294
+ -44.13833
+ -46.70085
+ -46.66149
+ -48.35883
+ -44.36377
+ -45.50144
+ -46.42204
+ -46.08328
+ -45.73318
+ -47.84658
+ -49.00233
+ -45.13598
+ -47.92345
+ -46.42874
+ -46.74604
+ -39.82155
+ -39.37421
+ -45.51609
+ -44.50432
+ -41.70202
+ -47.21475
+ -45.88815
+ -47.65540
+ -45.71832
+ -40.50336
+ -39.23597
+ -40.37950
+ -47.48698
+ -44.89493
+ -48.01861
+ -48.67346
+ -43.10340
+ -42.93694
+ -39.12726
+ -47.02554
