+++ sharc/propagation/P452/data_oxygen.txt
